Create a new conftest in subfolder
